Q2 built as a resolver module: an extern module alias is a module of the resolver's own under the canonical namespace <declaring module>::alias - written once by the import engine on the block (ExternBlockNode.binding_namespace, the derivation the type-declaration pass uses for the module's own key) and carried by every declaration the block brings in - its functions, the engine's types and its constants (declared at all for the first time, with the const-table key following) are declared and exported in that module by the name layer (set_module / restore_scope around the declarations; declare_c_imported_global, record_const_value_in; declare_c_imported_type and the whole-spelling symbol with no source module go), the alias symbol's qualified name IS that namespace, and every consumer reads Def(<mod>::alias) through the doors a graph module's members use: a scope segment bound to the alias - directly, or through an import of the exported alias (import compiler::codegen::{ llvm }), which alias_module_of follows as symbol_names_type does - is Def(<mod>::alias), so cit::add3() is resolve_module_qualified_symbol's under <mod>::cit::add3, cit::FLAG_A is scope_global_type's (leaf, ns) slot and codegen's resolve_global_in_namespace, cit::Mode::MODE_ON's segment is walk_module_rooted_type's Def(<mod>::cit::Mode), and a type spelling cit::Vec2 takes the same walk (the graph does not hold the alias, so modules_written_as cannot see it; the walk asks the alias module's export set for the leaf, one segment deep); the global scope's Namespace symbols are module identities under their full names and are not aliases; the extern prototypes register once, under <mod>::alias::fn, PUBLIC - the alias exists to be reached through and the name layer had declared them so while the index had them private, which the C-import door never asked and the module lane's visibility gate does (the LSP build refused 478 llvm::* calls until the two agreed); THE NAMESPACE IS PER DECLARING MODULE - two files each writing `extern module libc as cit;` get A::cit and B::cit, as two `extern crate libc as cit;` bind two names; a global cit would be first-claimant-wins again, do not fix it into one; the shape was put to Jake in section 8.257 and ruled by the brief's author for him to overrule (the resolver already owns one scope per module by namespace, nothing measured reads the graph for an alias, a file-less ModuleInfo row would have had 55 graph loops to skip), no graph loop touched; deleted with the alias::X key space: codegen's resolve_global ladder and its callers' spelling arm (a stamp naming no module names no global), the alias::name function registration codegen kept beside the C name (a pin resolves by the entry's symbol), the call resolver's try_resolve_cimport_function and the callee hint's bare_sym door, sema's alias_global, the index's global_vars (the bare-leaf last-write-wins map section 8.187's slot had superseded; with lookup_global gone it had no reader), register_global, lookup_global, get_global_module, get_global_namespace; the codegen global cache has one key, <namespace>::<leaf>, for a definition and a cross-module extern alike (SymbolResolver::global_cache_key_of); decl_global_key's binding branch is (leaf, binding namespace) like every global; scope_qualifier_type's Def arm reads through def_type's recording door now that no alias produces a type-relative stamp the index cannot serve; a finding the move surfaced: a C-import call was never argument-checked (the C-import door pinned the callee by spelling and returned its type) and through the module lane it is - tests/lang/c_import_libclang.cryo's probe::bindgen_probe_vfmt(buf, size, ...) passed a u64 to a C `unsigned long`, u32 on Windows by bindgen's target-dependent mapping, E0214 under the tree where it narrowed silently before; the fixture's parameter is `unsigned long long` (tests/helpers/bindgen_probe.h, abi_helpers.c), 64 bits on both hosts, the only change to what the va_list test pins; the compiler's own 478 llvm::* calls type-checked on the first try; measured (scripts/ns-migration/8.259/probe_alias.py over corpus2.sh W5p: lsp direct exit 0, failing halves 0, W5p-lines.txt 518 lines): ALIAS-SCOPE 506 (compiler::codegen::llvm 478 from the LSP half building codegen - section 8.257's 956 counted each call at two doors - FfiCImport::CImportTest::cit 17, CImportLibclang::probe 11), ALIAS-PATH 12 (section 8.257's 11 member-group type spellings and cit::Mode), ALIAS-MISS 0; objects (hash-tree.sh W6 against U3, 02ee243e's compiler): examples 0 of 1,126, tests 2 of 2,721 - FfiCImport/CImportTest.o and Lang/CImportLibclang.o, the two C-import users and nothing else, the mover named by nm on CImportTest.o: 15 symbols renamed, every one a C-import constant mangled under its module (C$3cit.6FLAG_A to C$10FfiCImport.11CImportTest.3cit.6FLAG_A), 33 unchanged; gates: lsp-check 264 modules 0 errors 487 warnings, cross-check 0 errors, test-census OK (PASS 65 projects 194 negatives, 2,127 unit, 68 pinned), ffi_c_import's 19 tests pass, lane-check predicted and re-pinned - LOOKUP 10 to 8 (the ladder's lookup_global, alias_global), LOOKUP_OTHER 95 to 92 (the ladder's three index reads), REGISTER 73 to 71 (the alias::name registration and its visibility), LOOKUP_ROUTED 53 to 51 (the hint's bare_sym door, sema's alias_global), LOOKUP_LOCAL 57 to 55 (the alias::name function-registry write, the ladder's cache read), DEFID_MINT 20 to 22 (the scope segment's alias answer and the walk's alias branch, a name becoming an answer at the name layer where every other mint is) - the gate's parser control refused the tree (`does not declare lookup_global as name-crossing`) until the deleted name left its LOOKUP list (scripts/lane-gate.py: three per-kind names; the self-test's fixture follows, 23 mutations behaved), ns-status-check OK 296 rows, verify-pin OK; new API for Jake: ExternBlockNode.binding_namespace, NameResolver::alias_module_of, declare_c_imported_global, record_const_value_in, SymbolResolver::global_cache_key_of; section 0: Q2's row LANDED (four checks), D12's LOOKUP_ROUTED 51, the try-ladder row (the C-import rung gone), the LOOKUP row 8 (three names), the set_decl_visibility check 4, the global's second-key row (the map deleted, two checks), the static-call-hint row's bare_sym check 0, resolve_path's lookup count 10 (section 8.259)

+++ tests/helpers/abi_helpers.c
@@ -178,6 +178,6 @@ int bindgen_probe_vstrlen(int n, va_list ap) {
 // Forward the SAME va_list a second time, into real libc `vsnprintf` - proves a
 // va_list survives a double hand-off (Cryo args -> this C fn -> libc) and that
 // format-string processing reads the forwarded list correctly.
-int bindgen_probe_vfmt(char *buf, unsigned long size, const char *fmt, va_list ap) {
-    return vsnprintf(buf, size, fmt, ap);
+int bindgen_probe_vfmt(char *buf, unsigned long long size, const char *fmt, va_list ap) {
+    return vsnprintf(buf, (size_t)size, fmt, ap);
 }
